Fix reward overflow bug
The reward was previously calculated as f64::from(next_score - score).
Since score and next_score are both of type u32, the result overflowed
whenever this difference was negative. Now the reward is calculated as
f64::from(next_score) - f64::from(score) to avoid overflows

+++ src/env_thread/env/episode/basic_episode/mod.rs
@@ -36,7 +36,10 @@ impl BasicEpisode {
         let reward = if terminated {
             0.0
         } else {
-            f64::from(next_score - score)
+            // IMPORTANT: The difference can be negative, so order of operations
+            // is critical here. The difference could overflow if it were
+            // calculated as f64::from(next_score - score)
+            f64::from(next_score) - f64::from(score)
         };
         transition_queue.push_back((
             (state_slice, next_state_slice, action, reward, terminated),
